+++ HIMSA Noahlink Wireless Checker.py
@@ -3,6 +3,9 @@
 #                   Copyright Bluebotlabz                   #
 #                                                           #
 #############################################################
+import os
+import feedparser
+from urllib.parse import urlparse
 from pathlib import Path
 from colorama import just_fix_windows_console
 from colorama import Fore, Back, Style
@@ -39,15 +42,28 @@ disclaimer = [
 if not turboFile.is_file():
     libhearingdownloader.printDisclaimer(disclaimer)
 
+# Get HIMSA Noahlink Wireless Download update using 3rd party web service (PolitePol)
+newsfeed = feedparser.parse('https://politepol.com/fd/smVib0t95riP')
+link0 = newsfeed.entries[0].link
+filename0 = os.path.basename(urlparse(link0).path)
+link1 = newsfeed.entries[1].link
+filename1 = os.path.basename(urlparse(link1).path)
 
 # Define list of valid versions and their download links
 validVersions = [
+    ("Current Downloads", "--"),
+    ("=================", "--"),
+    (newsfeed.entries[0].title, filename0, newsfeed.entries[0].link),
+    (newsfeed.entries[1].title, filename1, newsfeed.entries[1].link),
+    (" ", "--"),
+    ("Archived Downloads", "--"),
+    ("==================", "--"),
     ("Noahlink Wireless Driver v1.1.0.0", "for Windows 10 and 11 (March 2017)", "https://himsafiles.com/NoahlinkWireless/Driver_NLW_V.1.1.0.0.exe"),
     ("Noahlink Wireless Firmware v3.3.0.0", "v2.25 (NW1) / v3.23 (NW2) (April 2025)", "https://himsafiles.com/NoahlinkWireless/NLWUpgrader_3.3.0.0.exe"),
     ("Noahlink Wireless Firmware v3.1.0.92", "v2.25 (NW1) / v3.17 (NW2) (May 2024)", "https://himsafiles.com/NoahlinkWireless/NLWUpgrader_3.1.0.92.exe"),
     ("Noahlink Wireless Firmware v2.24", "v2.24 (NW1 only)", "https://himsafiles.com/NoahlinkWireless/NLWUpgrader2.24.exe"),
 ]
-print("\n\nThe latest available version is " + Fore.GREEN + "Driver v1.1.0.0" + Style.RESET_ALL + " / " + Fore.GREEN + "Firmware v3.3.0.0" + Style.RESET_ALL + "\n\n")
+print("\n\nThe latest available version is " + Fore.GREEN + newsfeed.entries[0].title + Style.RESET_ALL + " / " + Fore.GREEN + newsfeed.entries[1].title + Style.RESET_ALL + "\n\n")
 
 # Select outputDir and targetVersion
 outputDir = libhearingdownloader.selectOutputFolder()