+++ CHANGES.md
@@ -191,3 +191,6 @@
 - Phonak/Unitron/Hansaton/Oticon/Bernafon/Sonic/Philips: script tweak for easy changing target market and base version
 - Phonak/Unitron/Hansaton/Oticon/Bernafon/Sonic/Philips/Signia/Rexton/Audio Service/A&M: prompt target market (country) code before checking
 - Starkey: prompt current location (country name) before checking
+
+### Pre-release (work-in-progress)
+- HIMSA: checking actual update via 3rd-party web service